cleanup Dominion candidate data - issue #478
